@@ -12,22 +12,22 @@ class CacheItem implements CacheItemInterface
     /**
      * @var string
      */
-    protected $key;
+    protected string $key;
     
     /**
      * @var mixed
      */
-    protected $value;
+    protected mixed $value;
     
     /**
      * @var boolean
      */
-    protected $hit;
+    protected bool $hit;
     
     /**
      * @var DateTime
      */
-    protected $expiration;
+    protected DateTimeInterface $expiration;
 
     /**
      * CacheItem constructor.
@@ -35,7 +35,7 @@ class CacheItem implements CacheItemInterface
      * @param mixed $value
      * @param bool $hit
      */
-    public function __construct($key, $value, $hit = true)
+    public function __construct(string $key, mixed $value, bool $hit = true)
     {
         $this->key = $key;
         $this->value = $value;
@@ -46,7 +46,7 @@ class CacheItem implements CacheItemInterface
     /**
      * {@inheritdoc}
      */
-    public function getKey()
+    public function getKey(): string
     {
         return $this->key;
     }
@@ -54,14 +54,14 @@ class CacheItem implements CacheItemInterface
     /**
      * {@inheritdoc}
      */
-    public function get()
+    public function get(): mixed
     {
         return $this->isHit() ? $this->value : null;
     }
     /**
      * {@inheritdoc}
      */
-    public function set($value = null)
+    public function set(mixed $value = null): static
     {
         $this->value = $value;
         $this->hit = !is_null($value);
@@ -70,25 +70,27 @@ class CacheItem implements CacheItemInterface
     /**
      * {@inheritdoc}
      */
-    public function isHit()
+    public function isHit(): bool
     {
         return $this->hit;
     }
     /**
      * {@inheritdoc}
      */
-    public function expiresAt($expiration)
+    public function expiresAt(?DateTimeInterface $expiration): static
     {
-        $this->expiration = new DateTime('now +1 year');
-        if ($expiration instanceof DateTimeInterface) {
-            $this->expiration = $expiration;
+        if (empty($expiration)) {
+            $this->expiration = new DateTime('now +1 year');
+            return $this;
         }
+
+        $this->expiration = $expiration;
         return $this;
     }
     /**
      * {@inheritdoc}
      */
-    public function expiresAfter($time)
+    public function expiresAfter(int|\DateInterval|null $time): static
     {
         $this->expiration = new DateTime('now +1 year');
         if (is_numeric($time)) {
@@ -104,12 +106,12 @@ class CacheItem implements CacheItemInterface
     /**
      * @return DateTime
      */
-    public function getExpiresAt()
+    public function getExpiresAt(): DateTime
     {
         return $this->expiration;
     }
     
-    public function getExpiresInSecs()
+    public function getExpiresInSecs(): int
     {
         return $this->getExpiresAt()->getTimestamp() - (new DateTime('now'))->getTimestamp();
     }
