@@ -9,5 +9,5 @@ interface CacheAvailabilityInterface
      * Return if this CacheEngine is available for use
      * @return bool
      */
-    public function isAvailable();
+    public function isAvailable(): bool;
 }
