@@ -13,14 +13,14 @@ use ByJG\Cache\Psr6\CachePool;
 
 class Factory
 {
-    public static function createNullPool()
+    public static function createNullPool(): CachePool
     {
         return new CachePool(
             new NoCacheEngine()
         );
     }
 
-    public static function createSessionPool($prefix = null, $bufferSize = null)
+    public static function createSessionPool($prefix = null, $bufferSize = null): CachePool
     {
         return new CachePool(
             new SessionCacheEngine($prefix),
@@ -28,7 +28,7 @@ class Factory
         );
     }
 
-    public static function createFilePool($prefix = null, $path = null, $bufferSize = null, $logger = null)
+    public static function createFilePool($prefix = null, $path = null, $bufferSize = null, $logger = null): CachePool
     {
         return new CachePool(
             new FileSystemCacheEngine($prefix, $path, $logger),
@@ -36,7 +36,7 @@ class Factory
         );
     }
 
-    public static function createShmopPool($config = [], $bufferSize = null, $logger = null)
+    public static function createShmopPool($config = [], $bufferSize = null, $logger = null): CachePool
     {
         return new CachePool(
             new ShmopCacheEngine($config, $logger),
@@ -44,7 +44,7 @@ class Factory
         );
     }
 
-    public static function createArrayPool($bufferSize = null, $logger = null)
+    public static function createArrayPool($bufferSize = null, $logger = null): CachePool
     {
         return new CachePool(
             new ArrayCacheEngine($logger),
@@ -52,7 +52,7 @@ class Factory
         );
     }
 
-    public static function createMemcachedPool($servers = null, $bufferSize = null, $logger = null)
+    public static function createMemcachedPool($servers = null, $bufferSize = null, $logger = null): CachePool
     {
         return new CachePool(
             new MemcachedEngine($servers, $logger),
@@ -60,7 +60,7 @@ class Factory
         );
     }
 
-    public static function createRedisCacheEngine($servers = null, $password = null, $bufferSize = null, $logger = null)
+    public static function createRedisCacheEngine($servers = null, $password = null, $bufferSize = null, $logger = null): CachePool
     {
         return new CachePool(
             new RedisCacheEngine($servers, $password, $logger),
