@@ -32,7 +32,7 @@ class FileSystemCacheEngine extends BaseCacheEngine implements CacheLockInterfac
      * @return mixed Description
      * @throws \Psr\SimpleCache\InvalidArgumentException
      */
-    public function get($key, $default = null)
+    public function get(string $key, mixed $default = null): mixed
     {
         // Check if file is Locked
         $fileKey = $this->fixKey($key);
@@ -78,7 +78,7 @@ class FileSystemCacheEngine extends BaseCacheEngine implements CacheLockInterfac
      *
      *   MUST be thrown if the $key string is not a legal value.
      */
-    public function set($key, $value, $ttl = null)
+    public function set(string $key, mixed $value, DateInterval|int|null $ttl = null): bool
     {
         $fileKey = $this->fixKey($key);
 
@@ -119,7 +119,7 @@ class FileSystemCacheEngine extends BaseCacheEngine implements CacheLockInterfac
      * @return bool
      * @throws \Psr\SimpleCache\InvalidArgumentException
      */
-    public function delete($key)
+    public function delete(string $key): bool
     {
         $this->set($key, null);
         return true;
@@ -129,7 +129,7 @@ class FileSystemCacheEngine extends BaseCacheEngine implements CacheLockInterfac
      * Lock resource before set it.
      * @param string $key
      */
-    public function lock($key)
+    public function lock(string $key): void
     {
         $this->logger->info("[Filesystem cache] Lock '$key'");
 
@@ -146,7 +146,7 @@ class FileSystemCacheEngine extends BaseCacheEngine implements CacheLockInterfac
      * UnLock resource after set it.
      * @param string $key
      */
-    public function unlock($key)
+    public function unlock($key): void
     {
 
         $this->logger->info("[Filesystem cache] Unlock '$key'");
@@ -158,7 +158,7 @@ class FileSystemCacheEngine extends BaseCacheEngine implements CacheLockInterfac
         }
     }
 
-    public function isAvailable()
+    public function isAvailable(): bool
     {
         return is_writable(dirname($this->fixKey('test')));
     }
@@ -178,7 +178,7 @@ class FileSystemCacheEngine extends BaseCacheEngine implements CacheLockInterfac
      *
      * @return bool True on success and false on failure.
      */
-    public function clear()
+    public function clear(): bool
     {
         $patternKey = $this->fixKey('*');
         $list = glob($patternKey);
@@ -200,7 +200,7 @@ class FileSystemCacheEngine extends BaseCacheEngine implements CacheLockInterfac
      * @throws \Psr\SimpleCache\InvalidArgumentException
      *   MUST be thrown if the $key string is not a legal value.
      */
-    public function has($key)
+    public function has(string $key): bool
     {
         $fileKey = $this->fixKey($key);
         if (file_exists($fileKey)) {
