@@ -65,7 +65,7 @@ class MemcachedEngine extends BaseCacheEngine
      * @return mixed Description
      * @throws StorageErrorException
      */
-    public function get($key, $default = null)
+    public function get(string $key, mixed $default = null): mixed
     {
         $this->lazyLoadMemCachedServers();
 
@@ -85,7 +85,7 @@ class MemcachedEngine extends BaseCacheEngine
      * @return bool If the object is successfully posted
      * @throws StorageErrorException
      */
-    public function set($key, $value, $ttl = null)
+    public function set(string $key, mixed $value, DateInterval|int|null $ttl = null): bool
     {
         $this->lazyLoadMemCachedServers();
 
@@ -105,7 +105,7 @@ class MemcachedEngine extends BaseCacheEngine
      * @return bool
      * @throws StorageErrorException
      */
-    public function delete($key)
+    public function delete(string $key): bool
     {
         $this->lazyLoadMemCachedServers();
 
@@ -113,7 +113,7 @@ class MemcachedEngine extends BaseCacheEngine
         return true;
     }
 
-    public function isAvailable()
+    public function isAvailable(): bool
     {
         if (!class_exists('\Memcached')) {
             return false;
@@ -131,7 +131,7 @@ class MemcachedEngine extends BaseCacheEngine
      * @return bool
      * @throws StorageErrorException
      */
-    public function clear()
+    public function clear(): bool
     {
         $this->lazyLoadMemCachedServers();
         $result = $this->memCached->flush();
@@ -143,7 +143,7 @@ class MemcachedEngine extends BaseCacheEngine
      * @return bool
      * @throws StorageErrorException
      */
-    public function has($key)
+    public function has(string $key): bool
     {
         $this->lazyLoadMemCachedServers();
 
