@@ -4,6 +4,7 @@ namespace ByJG\Cache\Psr16;
 
 use ByJG\Cache\Exception\InvalidArgumentException;
 use ByJG\Cache\Exception\StorageErrorException;
+use DateInterval;
 use Psr\Log\NullLogger;
 
 /**
@@ -73,7 +74,7 @@ class ShmopCacheEngine extends BaseCacheEngine
      * @param mixed $default The time to live in seconds of the object. Depends on implementation.
      * @return mixed The Object
      */
-    public function get($key, $default = null)
+    public function get(string $key, mixed $default = null): mixed
     {
        if ($default === false) {
             $this->logger->info("[Shmop Cache] Ignored  $key because TTL=FALSE");
@@ -129,7 +130,7 @@ class ShmopCacheEngine extends BaseCacheEngine
      * @throws InvalidArgumentException
      * @throws StorageErrorException
      */
-    public function set($key, $value, $ttl = null)
+    public function set(string $key, mixed $value, DateInterval|int|null $ttl = null): bool
     {
         $this->logger->info("[Shmop Cache] set '$key'");
 
@@ -173,7 +174,7 @@ class ShmopCacheEngine extends BaseCacheEngine
      * @param string $key
      * @return bool
      */
-    public function delete($key)
+    public function delete(string $key): bool
     {
         $this->logger->info("[Shmop Cache] release '$key'");
 
@@ -208,16 +209,17 @@ class ShmopCacheEngine extends BaseCacheEngine
         }
     }
 
-    public function clear()
+    public function clear(): bool
     {
         $patternKey = sys_get_temp_dir() . '/shmop-*.cache';
         $list = glob($patternKey);
         foreach ($list as $file) {
             $this->deleteFromFilenameToken($file);
         }
+        return true;
     }
 
-    public function has($key)
+    public function has(string $key): bool
     {
         $file = $this->getFilenameToken($key);
         $fileKey = $this->getFTok($file);
@@ -236,7 +238,7 @@ class ShmopCacheEngine extends BaseCacheEngine
     }
 
 
-    public function isAvailable()
+    public function isAvailable(): bool
     {
         return function_exists('shmop_open');
     }
